requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
supabase==2.4.3
python-dotenv==1.0.1